feat: migrate all service registry state file to api primitives.rs
