+++ ampd/src/commands/deregister_chain_support.rs
@@ -4,7 +4,7 @@ use cosmrs::tx::Msg;
 use error_stack::Result;
 use report::ResultCompatExt;
 use router_api::ChainName;
-use service_registry::msg::ExecuteMsg;
+use service_registry_api::msg::ExecuteMsg;
 use valuable::Valuable;
 
 use crate::commands::{broadcast_tx, verifier_pub_key};
