+++ ampd/src/commands/unbond_verifier.rs
@@ -3,7 +3,7 @@ use cosmrs::cosmwasm::MsgExecuteContract;
 use cosmrs::tx::Msg;
 use error_stack::Result;
 use report::ResultCompatExt;
-use service_registry::msg::ExecuteMsg;
+use service_registry_api::msg::ExecuteMsg;
 use valuable::Valuable;
 
 use crate::commands::{broadcast_tx, verifier_pub_key};
