+++ contracts/multisig-prover/Cargo.toml
@@ -54,7 +54,7 @@ multisig = { workspace = true, features = ["library"] }
 report = { workspace = true }
 router-api = { workspace = true }
 serde_json = "1.0.89"
-service-registry = { workspace = true }
+service-registry-api = { workspace = true }
 sha3 = { workspace = true }
 stellar = { workspace = true }
 sui-gateway = { workspace = true }
