+++ contracts/service-registry/src/contract.rs
@@ -7,10 +7,11 @@ use cosmwasm_std::{
 };
 use error_stack::{bail, Report, ResultExt};
 
-use crate::error::ContractError;
+// use crate::error::ContractError;
 use crate::msg::InstantiateMsg;
-use crate::state::{AuthorizationState, BondingState, Service, SERVICES, VERIFIERS};
+use service_registry_api::error::ContractError;
 use service_registry_api::msg::{ExecuteMsg, QueryMsg};
+use service_registry_api::{AuthorizationState, BondingState, Service, SERVICES, VERIFIERS};
 
 mod execute;
 mod migrations;
@@ -195,8 +196,8 @@ mod test {
     use router_api::ChainName;
 
     use super::*;
-    use crate::msg::VerifierDetails;
-    use crate::state::{Verifier, WeightedVerifier, VERIFIER_WEIGHT};
+    use service_registry_api::msg::VerifierDetails;
+    use service_registry_api::{Verifier, WeightedVerifier, VERIFIER_WEIGHT};
 
     const GOVERNANCE_ADDRESS: &str = "governance";
     const UNAUTHORIZED_ADDRESS: &str = "unauthorized";
