+++ contracts/service-registry/src/lib.rs
@@ -1,12 +1,9 @@
 pub mod client;
 pub mod contract;
-pub mod error;
 pub mod helpers;
 pub mod msg;
 mod state;
 
-pub use state::{
+pub use service_registry_api::{
     AuthorizationState, BondingState, Service, Verifier, WeightedVerifier, VERIFIER_WEIGHT,
 };
-
-pub use crate::error::ContractError;
