+++ integration-tests/Cargo.toml
@@ -46,6 +46,7 @@ router-api = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 service-registry = { workspace = true }
+service-registry-api = { workspace = true }
 sha3 = { workspace = true }
 tofn = { workspace = true }
 voting-verifier = { workspace = true }
