+++ integration-tests/tests/test_utils/mod.rs
@@ -24,7 +24,7 @@ use multisig::verifier_set::VerifierSet;
 use multisig_prover::msg::VerifierSetResponse;
 use rewards::PoolId;
 use router_api::{Address, ChainName, CrossChainId, GatewayDirection, Message};
-use service_registry::msg::ExecuteMsg;
+use service_registry_api::msg::ExecuteMsg;
 use sha3::{Digest, Keccak256};
 use tofn::ecdsa::KeyPair;
