+++ integration-tests/tests/update_worker_set.rs
@@ -2,8 +2,8 @@ use cosmwasm_std::Addr;
 use cw_multi_test::Executor;
 use integration_tests::contract::Contract;
 use multisig_prover::msg::ExecuteMsg;
-use service_registry::msg::QueryMsg as ServiceRegistryQueryMsg;
 use service_registry::WeightedVerifier;
+use service_registry_api::msg::QueryMsg as ServiceRegistryQueryMsg;
 
 pub mod test_utils;
