+++ packages/service-registry-api/Cargo.toml
@@ -9,12 +9,14 @@ edition = { workspace = true }
 axelar-wasm-std = { workspace = true, features = ["derive"] }
 cosmwasm-schema = { workspace = true }
 cosmwasm-std = { workspace = true }
+cw-storage-plus = { workspace = true }
 error-stack = { workspace = true }
 msgs-derive = { workspace = true }
 report = { workspace = true }
 router-api = { workspace = true }
 schemars = "0.8.10"
 serde = { version = "1.0.145", default-features = false, features = ["derive"] }
+thiserror = { workspace = true }
 
 [dev-dependencies]
 goldie = { workspace = true }
