+++ packages/service-registry-api/src/lib.rs
@@ -1,1 +1,6 @@
+mod primitives;
+pub mod error;
+
 pub mod msg;
+
+pub use primitives::*;
